adding rizzino_a06c.js
Having problems with getting my lottery program to run correctly.  I have tried coding it multiple different ways while trying to follow the instructions and have not been able to get a working result.  I have got the do while loop to load my array with numbers and created a for loop to iterate through the numbers but it will not return the matched numbers or winnings when the numbers match.  I am turning this in so it is not late but would appreciate any suggestions or help that could help me get it running correctly.