adding rizzino_a06c.js
Have fixed the final part of my switch structure to make it a if...else statement like you suggested.  I have readjusted my for loop to properly iterate though the lottoPicks array.  However I still have one issue of concern, while I have been able to use the switch structure to match the powerBallPick and return the result of winnings of $100, when the other winning numbers show up in the array it is not showing that it has a match or returning any winnings of $10.  I have tried to switch the placement of my loops and structure and which variable was being used to see if it gave me the result that I am looking for but still have had no luck.  If you have any other suggestions that could assist me in getting the program to run correctly I would greatly appreciate it.  Thank you.