adding rizzino_a06c.js
Fixed the remaining issue in the switch structure.  My task 2 program is now up and running! Thank you for the guidance I think I understand where I went wrong with the n variable now.  I have also removed the console.log for the lottoPicks array to improve efficiency and make the output of the program match figure 9 in the instructions.